pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^3.0.3"
python-dotenv = "^0.14.0"
requests = "^2.31.0"
pytest = "^8.1.1"
pymongo = "^4.10.1"
mongomock = "^4.2.0.post1"
flask-dance = "^7.1.0"
werkzeug = "3.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
